refactor: Use barrel file for db imports
