+++ src/db/index.ts
@@ -1,0 +1,3 @@
+import { type QueryClient, db, transaction } from './db'
+
+export { type QueryClient, db, transaction }
