+++ src/modules/user/userService.ts
@@ -1,12 +1,10 @@
-import type { QueryClient } from '@/db/db'
-import type { User } from '@/db/schema/user'
+import type { QueryClient } from '@/db'
 import { UserNotFoundException } from '@/modules/user/userExceptions'
 import * as userRepo from '@/modules/user/userRepo'
 
-export const findUser = (id: number): Promise<User | null> =>
-  userRepo.findUser(id)
+export const findUser = (id: number) => userRepo.findUser(id)
 
-export const getUser = async (id: number): Promise<User> => {
+export const getUser = async (id: number) => {
   const user = await findUser(id)
   if (user === null)
     throw new UserNotFoundException(`User with id ${id} not found`)
@@ -19,4 +17,4 @@ export const registerUser = (
     firstName: string
   },
   client: QueryClient,
-): Promise<number> => userRepo.createUser(user, client)
+) => userRepo.createUser(user, client)
